requirements.txt
dialog_bot_sdk
prometheus-client
protobuf
google
google-api-python-client
googleapis-common-protos
gevent
grpcio
grpcio-tools
requests
pyopenssl
Pillow